task_management/requirements.txt
asgiref==3.8.1
attrs==23.2.0
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
